# swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/app":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/api/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginUserDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"Login with mobile and password","tags":["auth"]}},"/api/auth/me":{"get":{"operationId":"AuthController_getProfile","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"پروفایل","tags":["auth"]}},"/api/auth/otp/request":{"post":{"operationId":"AuthController_requestOtp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestOtpDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"Request OTP code","tags":["auth"]}},"/api/auth/otp/verify":{"post":{"operationId":"AuthController_verifyOtp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyOtpDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"Verify OTP and login","tags":["auth"]}},"/api/auth/pre-register":{"post":{"operationId":"AuthController_preRegister","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PreRegisterDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"Start registration process and send OTP","tags":["auth"]}},"/api/auth/verify-register":{"post":{"operationId":"AuthController_verifyAndRegister","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyOtpDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"Verify OTP and complete registration","tags":["auth"]}},"/api/auth/resend-register-otp":{"post":{"operationId":"AuthController_resendRegisterOtp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestOtpDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"Resend registration OTP","tags":["auth"]}},"/api/movies":{"get":{"description":"دریافت لیست صفحه‌بندی شده فیلم‌ها","operationId":"MovieController_findAll","parameters":[{"name":"search","required":false,"in":"query","description":"جستجو در عنوان","schema":{"type":"string"}},{"name":"genre","required":false,"in":"query","description":"فیلتر بر اساس ژانر","schema":{"type":"number"}},{"name":"year","required":false,"in":"query","description":"فیلتر بر اساس سال","schema":{"type":"number"}},{"name":"sortBy","required":false,"in":"query","description":"مرتب‌سازی","schema":{"enum":["popularity","voteAverage","releaseDate"],"type":"string"}},{"name":"sortOrder","required":false,"in":"query","description":"ترتیب مرتب‌سازی","schema":{"enum":["asc","desc"],"type":"string"}},{"name":"page","required":false,"in":"query","description":"شماره صفحه","schema":{"default":1,"type":"number"}},{"name":"limit","required":false,"in":"query","description":"تعداد آیتم در هر صفحه","schema":{"default":20,"type":"number"}}],"responses":{"200":{"description":"لیست فیلم‌ها","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedResponseDto"}}}}},"summary":"دریافت لیست فیلم‌ها","tags":["فیلم‌ها"]}},"/api/movies/{id}":{"get":{"description":"دریافت اطلاعات کامل یک فیلم","operationId":"MovieController_findOne","parameters":[{"name":"id","required":true,"in":"path","description":"شناسه فیلم","schema":{"type":"number"}}],"responses":{"200":{"description":"جزئیات فیلم","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MovieDetailDto"}}}},"404":{"description":"فیلم یافت نشد"}},"summary":"دریافت جزئیات فیلم","tags":["فیلم‌ها"]}},"/api/movies/{id}/credits":{"get":{"description":"دریافت لیست بازیگران و عوامل فیلم","operationId":"MovieController_findCredits","parameters":[{"name":"id","required":true,"in":"path","description":"شناسه فیلم","schema":{"type":"number"}}],"responses":{"200":{"description":"بازیگران فیلم","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MovieCreditsDto"}}}},"404":{"description":"فیلم یافت نشد"}},"summary":"دریافت بازیگران فیلم","tags":["فیلم‌ها"]}},"/api/series":{"get":{"description":"دریافت لیست صفحه‌بندی شده سریال‌ها","operationId":"SeriesController_findAll","parameters":[{"name":"search","required":false,"in":"query","description":"جستجو در عنوان","schema":{"type":"string"}},{"name":"genre","required":false,"in":"query","description":"فیلتر بر اساس ژانر","schema":{"type":"number"}},{"name":"year","required":false,"in":"query","description":"فیلتر بر اساس سال","schema":{"type":"number"}},{"name":"sortBy","required":false,"in":"query","description":"مرتب‌سازی","schema":{"enum":["popularity","voteAverage","firstAirDate"],"type":"string"}},{"name":"sortOrder","required":false,"in":"query","description":"ترتیب مرتب‌سازی","schema":{"enum":["asc","desc"],"type":"string"}},{"name":"page","required":false,"in":"query","description":"شماره صفحه","schema":{"default":1,"type":"number"}},{"name":"limit","required":false,"in":"query","description":"تعداد آیتم در هر صفحه","schema":{"default":20,"type":"number"}}],"responses":{"200":{"description":"لیست سریال‌ها","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedResponseDto"}}}}},"summary":"دریافت لیست سریال‌ها","tags":["سریال‌ها"]}},"/api/series/{id}":{"get":{"description":"دریافت اطلاعات کامل یک سریال","operationId":"SeriesController_findOne","parameters":[{"name":"id","required":true,"in":"path","description":"شناسه سریال","schema":{"type":"number"}}],"responses":{"200":{"description":"جزئیات سریال","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SeriesDetailDto"}}}},"404":{"description":"سریال یافت نشد"}},"summary":"دریافت جزئیات سریال","tags":["سریال‌ها"]}},"/api/series/{id}/seasons":{"get":{"description":"دریافت لیست تمام فصل‌های یک سریال","operationId":"SeriesController_findSeasons","parameters":[{"name":"id","required":true,"in":"path","description":"شناسه سریال","schema":{"type":"number"}}],"responses":{"200":{"description":"فصل‌های سریال","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SeasonDto"}}}}},"404":{"description":"سریال یافت نشد"}},"summary":"دریافت فصل‌های سریال","tags":["سریال‌ها"]}},"/api/series/{id}/season/{seasonNumber}":{"get":{"description":"دریافت اطلاعات یک فصل خاص از سریال","operationId":"SeriesController_findSeason","parameters":[{"name":"id","required":true,"in":"path","description":"شناسه سریال","schema":{"type":"number"}},{"name":"seasonNumber","required":true,"in":"path","description":"شماره فصل","schema":{"type":"number"}}],"responses":{"200":{"description":"جزئیات فصل","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SeasonDto"}}}},"404":{"description":"فصل یافت نشد"}},"summary":"دریافت فصل خاص","tags":["سریال‌ها"]}},"/api/people":{"get":{"description":"دریافت لیست صفحه‌بندی شده اشخاص","operationId":"PersonController_findAll","parameters":[{"name":"search","required":false,"in":"query","description":"جستجو در نام","schema":{"type":"string"}},{"name":"sortBy","required":false,"in":"query","description":"مرتب‌سازی","schema":{"enum":["popularity","name"],"type":"string"}},{"name":"sortOrder","required":false,"in":"query","description":"ترتیب مرتب‌سازی","schema":{"enum":["asc","desc"],"type":"string"}},{"name":"page","required":false,"in":"query","description":"شماره صفحه","schema":{"default":1,"type":"number"}},{"name":"limit","required":false,"in":"query","description":"تعداد آیتم در هر صفحه","schema":{"default":20,"type":"number"}}],"responses":{"200":{"description":"لیست اشخاص","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedResponseDto"}}}}},"summary":"دریافت لیست اشخاص","tags":["اشخاص"]}},"/api/people/{id}":{"get":{"description":"دریافت اطلاعات کامل یک شخص","operationId":"PersonController_findOne","parameters":[{"name":"id","required":true,"in":"path","description":"شناسه شخص","schema":{"type":"number"}}],"responses":{"200":{"description":"جزئیات شخص","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PersonDetailDto"}}}},"404":{"description":"شخص یافت نشد"}},"summary":"دریافت جزئیات شخص","tags":["اشخاص"]}},"/api/search":{"get":{"description":"جستجو در فیلم‌ها، سریال‌ها و اشخاص","operationId":"SearchController_search","parameters":[{"name":"q","required":true,"in":"query","description":"کلمه کلیدی جستجو","schema":{"type":"string"}},{"name":"type","required":false,"in":"query","description":"نوع جستجو","schema":{"enum":["movie","series","person"],"type":"string"}}],"responses":{"200":{"description":"نتایج جستجو","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SearchResultDto"}}}}}},"summary":"جستجو","tags":["جستجو"]}}},"info":{"title":"Backend Api","description":"","version":"1.0","contact":{}},"tags":[{"name":"","description":""}],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"LoginUserDto":{"type":"object","properties":{"mobile":{"type":"string"},"password":{"type":"string"}},"required":["mobile","password"]},"RequestOtpDto":{"type":"object","properties":{"mobile":{"type":"string"}},"required":["mobile"]},"VerifyOtpDto":{"type":"object","properties":{"mobile":{"type":"string"},"otp":{"type":"string"}},"required":["mobile","otp"]},"PreRegisterDto":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"password":{"type":"string"},"mobile":{"type":"string"},"nationalCode":{"type":"string"},"birthDate":{"type":"string"}},"required":["firstName","lastName","password","mobile","nationalCode","birthDate"]},"PaginatedResponseDto":{"type":"object","properties":{"page":{"type":"number","description":"شماره صفحه"},"totalResults":{"type":"number","description":"تعداد کل نتایج"},"totalPages":{"type":"number","description":"تعداد کل صفحات"},"results":{"description":"نتایج","type":"array","items":{"type":"string"}}},"required":["page","totalResults","totalPages","results"]},"MovieDetailDto":{"type":"object","properties":{"id":{"type":"number","description":"شناسه فیلم"},"title":{"type":"string","description":"عنوان فیلم"},"description":{"type":"string","description":"توضیحات فیلم"},"releaseDate":{"format":"date-time","type":"string","description":"تاریخ انتشار"},"voteAverage":{"type":"number","description":"امتیاز متوسط"},"posterPath":{"type":"string","description":"مسیر پوستر"},"backdropPath":{"type":"string","description":"مسیر تصویر پس‌زمینه"},"genres":{"description":"ژانرها","type":"array","items":{"type":"string"}},"runtime":{"type":"number","description":"مدت زمان فیلم (دقیقه)"},"status":{"type":"string","description":"وضعیت"},"originalLanguage":{"type":"string","description":"زبان اصلی"},"budget":{"type":"number","description":"بودجه"},"revenue":{"type":"number","description":"درآمد"},"voteCount":{"type":"number","description":"تعداد رای‌ها"},"createdAt":{"format":"date-time","type":"string","description":"تاریخ ایجاد"},"updatedAt":{"format":"date-time","type":"string","description":"تاریخ بروزرسانی"}},"required":["id","title","description","releaseDate","voteAverage","posterPath","backdropPath","genres","runtime","status","originalLanguage","budget","revenue","voteCount","createdAt","updatedAt"]},"MovieCreditsDto":{"type":"object","properties":{"id":{"type":"number","description":"شناسه فیلم"},"cast":{"description":"بازیگران","type":"array","items":{"type":"string"}}},"required":["id","cast"]},"SeriesDetailDto":{"type":"object","properties":{"id":{"type":"number","description":"شناسه سریال"},"title":{"type":"string","description":"عنوان سریال"},"description":{"type":"string","description":"توضیحات سریال"},"firstAirDate":{"format":"date-time","type":"string","description":"تاریخ انتشار اولین قسمت"},"voteAverage":{"type":"number","description":"امتیاز متوسط"},"posterPath":{"type":"string","description":"مسیر پوستر"},"backdropPath":{"type":"string","description":"مسیر تصویر پس‌زمینه"},"numberOfSeasons":{"type":"number","description":"تعداد فصل‌ها"},"numberOfEpisodes":{"type":"number","description":"تعداد قسمت‌ها"},"genres":{"description":"ژانرها","type":"array","items":{"type":"string"}},"lastAirDate":{"format":"date-time","type":"string","description":"تاریخ انتشار آخرین قسمت"},"status":{"type":"string","description":"وضعیت"},"originalLanguage":{"type":"string","description":"زبان اصلی"},"voteCount":{"type":"number","description":"تعداد رای‌ها"},"popularity":{"type":"number","description":"مشهوریت"},"createdAt":{"format":"date-time","type":"string","description":"تاریخ ایجاد"},"updatedAt":{"format":"date-time","type":"string","description":"تاریخ بروزرسانی"}},"required":["id","title","description","firstAirDate","voteAverage","posterPath","backdropPath","numberOfSeasons","numberOfEpisodes","genres","lastAirDate","status","originalLanguage","voteCount","popularity","createdAt","updatedAt"]},"SeasonDto":{"type":"object","properties":{"id":{"type":"number","description":"شناسه فصل"},"seasonNumber":{"type":"number","description":"شماره فصل"},"title":{"type":"string","description":"عنوان فصل"},"description":{"type":"string","description":"توضیحات فصل"},"airDate":{"format":"date-time","type":"string","description":"تاریخ انتشار"},"posterPath":{"type":"string","description":"مسیر پوستر"},"episodeCount":{"type":"number","description":"تعداد قسمت‌ها"},"episodes":{"description":"قسمت‌ها","type":"array","items":{"type":"string"}}},"required":["id","seasonNumber","title","description","airDate","posterPath","episodeCount","episodes"]},"PersonDetailDto":{"type":"object","properties":{"id":{"type":"number","description":"شناسه شخص"},"name":{"type":"string","description":"نام"},"profilePath":{"type":"string","description":"مسیر تصویر پروفایل"},"popularity":{"type":"number","description":"مشهوریت"},"alsoKnownAs":{"type":"string","description":"نام مستعار"},"birthday":{"format":"date-time","type":"string","description":"تاریخ تولد"},"placeOfBirth":{"type":"string","description":"محل تولد"},"biography":{"type":"string","description":"بیوگرافی"},"gender":{"type":"number","description":"جنسیت"},"status":{"type":"string","description":"وضعیت"},"createdAt":{"format":"date-time","type":"string","description":"تاریخ ایجاد"},"updatedAt":{"format":"date-time","type":"string","description":"تاریخ بروزرسانی"}},"required":["id","name","profilePath","popularity","alsoKnownAs","birthday","placeOfBirth","biography","gender","status","createdAt","updatedAt"]},"SearchResultDto":{"type":"object","properties":{"type":{"type":"string","description":"نوع نتیجه"},"results":{"description":"نتایج","type":"array","items":{"type":"string"}}},"required":["type","results"]}}}}